KNN_Weather_Predictor.ipynb
This KNN algorithm takes MinTemp, MaxTemp, and AverageTemp and makes an accurate prediction on the precipitation level. It also does the same for WindSpeed and WindDirection. I then compared the two to see which one was more accurate than the other if someone was to use this algorithm in the future to predict weather patterns.